+++ CHANGELOG.md
@@ -13,6 +13,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Parallelize variant loading for each chromosome
 ### Added
 - Shortcut button for HPO panel MEI variants from case page
+- Export managed variants from CLI
 
 ## [4.72.4]
 ### Changed